src/images/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.12.4
googleapis-common-protos==1.60.0
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
img2vec-pytorch==1.0.1
Jinja2==3.1.2
loguru==0.7.2
lz4==4.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
pika==1.3.2
Pillow==10.0.1
pinecone-client==2.2.4
portalocker==2.8.2
protobuf==4.24.3
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
PyYAML==6.0.1
qdrant-client==1.5.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
validators==0.22.0
weaviate-client==3.24.1
